Fix: 이미지 있는 뉴스만 배너에 띄우기
