+++ app/src/main/java/com/protect/jikigo/ui/home/news/NewsBesidesFragment.kt
@@ -68,15 +68,13 @@ class NewsBesidesFragment : Fragment() {
         viewModel.newsList.observe(viewLifecycleOwner) { newsList ->
             newsAdapter.submitList(newsList as MutableList<NewsItem>?)
 
-            val bannerItems = listOfNotNull(
-                newsList.getOrNull(2),
-                newsList.getOrNull(8),
-                newsList.getOrNull(14)
-            )
-            newsBannerAdapter.submitList(bannerItems as MutableList<NewsItem>?)
+            // 이미지가 있는 뉴스 아이템만 필터링
+            val newsWithImages = newsList.filter { it.imageUrl != null }
 
+            // 이미지가 있는 뉴스 아이템이 3개 이상이면 랜덤으로 3개 선택
+                val randomItems = newsWithImages.shuffled().take(3)
+                newsBannerAdapter.submitList(randomItems as MutableList<NewsItem>?)
         }
-
 
         category?.let { viewModel.fetchNews(it) }
     }
